services/secrets-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
cryptography==41.0.7
pyjwt==2.8.0
requests==2.31.0
python-multipart==0.0.6